fix(native): tell absent from empty and false in the ported jq rules
A Go zero value collapses three states jq keeps apart, and every prompt
divergence found in review came from that. The fields the shell
interpolates are now held as the JSON they arrived as.

- A comment left by a deleted account arrives as `author: null`, and
  json.Unmarshal of `null` into a string is a nil error that leaves the
  string empty, so the prompt said `- ****:` where jq says `- **null**:`.
- `//` falls through on null and false and prints an empty string as
  itself, so `why: ""` and `prior_resolution: ""` took defaults the shell
  does not take, and `min_fix_severity: ""` printed **medium**.
- `select(.isResolved == false)` is strict, so a thread with no
  isResolved is dropped; a Go bool showed it.
- A line number is whatever the payload wrote. lib/validate.sh checks
  only that it is a number, so 1.5 validates, renders `a.ts:1.5` and
  failed to decode into an int at all.
- Meta reaches the shell through command substitution, which strips
  trailing newlines. Render strips them rather than leaving a contract
  for the next caller, which builds Meta from `gh pr view --json body`.

The validator's numbers now carry their literal, their double and their
exact value, because jq uses a different one for each: `floor` is double
arithmetic, so 1.0000000000000001 and 1E+19 are whole, while `>` and `-`
compare at full precision, so both are still out of range. The earlier
`n != float64(int64(n))` was implementation-defined on overflow and
saturated at 2^63, reporting a semantic failure as a shape one — and the
two carry different retry budgets. It also refuses a lone surrogate
escape, which jq will not parse and Go silently replaced.

The embedded schemas and skills are unexported behind accessors that
answer a copy, and SkillBody clones, so no package can write through
process-wide rubric or schema bytes. The drift gate reads the
`//go:embed` directives rather than a hand-kept list, and refuses an
embed with no canonical source.

Tests: the commit-convention section is compared against lib/prompt.sh
run live over a real repository across ten cases, because no frozen
vector reaches it; 49 mutations that every earlier test survived are
each refused by a named case.

+++ internal/validate/resolve_test.go
@@ -8,7 +8,7 @@ import (
 
 // threeFindings is what tests/test-schemas.sh:155 supplies: three findings were
 // numbered, and two issues were offered as duplicate candidates.
-var threeFindings = &validate.Expectations{Findings: 3, Candidates: []float64{19, 31}}
+var threeFindings = &validate.Expectations{Findings: 3, Candidates: []int{19, 31}}
 
 // resolutions wraps a resolutions array the way tests/test-schemas.sh:157-160
 // does.
